Basic refactor for domain driven development. Separate application layers to api-direct, api-service, application, domain and infrastructure modules with central domain module as business model and logic with no dependencies. Commit before full tests written
