+++ app/api-direct/src/main/java/com/programmersonly/mentorship/direct/template/ConfirmTemplateRequest.java
@@ -1,0 +1,10 @@
+package com.programmersonly.mentorship.direct.template;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+class ConfirmTemplateRequest {
+  private UUID templateId;
+}
