+++ app/api-direct/src/main/java/com/programmersonly/mentorship/direct/template/CreateTemplateRequest.java
@@ -1,0 +1,12 @@
+package com.programmersonly.mentorship.direct.template;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+class CreateTemplateRequest {
+  private UUID userId;
+  private String email;
+  private String fullName;
+}
