+++ app/domain/src/main/java/com/programmersonly/mentorship/exception/ErrorResponse.java
@@ -1,0 +1,6 @@
+package com.programmersonly.mentorship.exception;
+
+public interface ErrorResponse {
+    ErrorCode getErrorCode();
+    String getMessage();
+}
