+++ app/domain/src/main/java/com/programmersonly/mentorship/exception/TemplateNotFoundException.java
@@ -1,0 +1,8 @@
+package com.programmersonly.mentorship.exception;
+
+public class TemplateNotFoundException extends BusinessException {
+
+    public TemplateNotFoundException() {
+        super(ErrorCode.MS02);
+    }
+}
