+++ app/domain/src/main/java/com/programmersonly/mentorship/template/TemplateService.java
@@ -1,0 +1,11 @@
+package com.programmersonly.mentorship.template;
+
+import com.programmersonly.mentorship.template.dto.CreateTemplateDto;
+
+import java.util.UUID;
+
+public interface TemplateService {
+    void create(CreateTemplateDto createTemplateDto);
+    void confirm(UUID templateId);
+    void remove(UUID templateId);
+}
