+++ app/domain/src/main/java/com/programmersonly/mentorship/template/TemplateStatus.java
@@ -1,0 +1,5 @@
+package com.programmersonly.mentorship.template;
+
+enum TemplateStatus {
+    CREATED, CONFIRMED
+}
