--- src/main/java/com/programmersonly/mentorship/commons/exception/BasicErrorResponse.java
@@ -1,20 +1,0 @@
-package com.programmersonly.mentorship.commons.exception;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class BasicErrorResponse implements ErrorResponse {
-
-    private final String errorCode;
-    private final String message;
-
-    @Override
-    public String getErrorCode() {
-        return this.errorCode;
-    }
-
-    @Override
-    public String getMessage() {
-        return this.message;
-    }
-}
