--- src/main/java/com/programmersonly/mentorship/commons/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package com.programmersonly.mentorship.commons.exception;
-
-public interface ErrorResponse {
-    String getErrorCode();
-    String getMessage();
-}
