--- src/main/java/com/programmersonly/mentorship/mentors/template/ConfirmTemplateRequest.java
@@ -1,10 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-class ConfirmTemplateRequest {
-  private UUID templateId;
-}
