--- src/main/java/com/programmersonly/mentorship/mentors/template/RemoveTemplateRequest.java
@@ -1,10 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-class RemoveTemplateRequest {
-  private UUID templateId;
-}
