--- src/main/java/com/programmersonly/mentorship/mentors/template/SpringTemplateCommand.java
@@ -1,20 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-class SpringTemplateCommand implements TemplateCommand {
-
-    private final TemplateEntityRepository repository;
-
-    @Override
-    public void save(TemplateEntity templateEntity) {
-        repository.save(templateEntity);
-    }
-
-    @Override
-    public void deleteById(UUID templateId) {
-        repository.deleteById(templateId);
-    }
-}
