--- src/main/java/com/programmersonly/mentorship/mentors/template/TemplateCommand.java
@@ -1,8 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import java.util.UUID;
-
-interface TemplateCommand {
-    void save(TemplateEntity templateEntity);
-    void deleteById(UUID templateId);
-}
