--- src/main/java/com/programmersonly/mentorship/mentors/template/TemplateController.java
@@ -1,37 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import lombok.AllArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@AllArgsConstructor
-@RequestMapping(value = "/template")
-class TemplateController {
-
-    private final TemplateFacade facade;
-
-    @PostMapping("/create")
-    public ResponseEntity<Void> createMentorTemplate(@RequestBody CreateTemplateRequest dto){
-        facade.create(dto);
-        return new ResponseEntity<>(HttpStatus.CREATED);
-    }
-
-    @PutMapping("/confirm")
-    public ResponseEntity<Void> confirmMentorTemplate(@RequestBody ConfirmTemplateRequest dto){
-        facade.confirm(dto);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-
-    @DeleteMapping("/delete")
-    public ResponseEntity<Void> removeMentorTemplate(@RequestBody RemoveTemplateRequest dto){
-        facade.remove(dto);
-        return new ResponseEntity<>(HttpStatus.OK);
-    }
-}
