--- src/main/java/com/programmersonly/mentorship/mentors/template/TemplateEntityRepository.java
@@ -1,8 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import java.util.UUID;
-import org.springframework.data.repository.CrudRepository;
-
-interface TemplateEntityRepository extends CrudRepository<TemplateEntity, UUID> {
-
-}
