--- src/main/java/com/programmersonly/mentorship/mentors/template/TemplateQuery.java
@@ -1,8 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-import java.util.Optional;
-import java.util.UUID;
-
-interface TemplateQuery {
-    Optional<TemplateEntity> findById(UUID templateId);
-}
