--- src/main/java/com/programmersonly/mentorship/mentors/template/TemplateService.java
@@ -1,7 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-interface TemplateService {
-    void create(CreateTemplateRequest request);
-    void confirm(ConfirmTemplateRequest request);
-    void remove(RemoveTemplateRequest request);
-}
