--- src/main/java/com/programmersonly/mentorship/mentors/template/TemplateStatus.java
@@ -1,5 +1,0 @@
-package com.programmersonly.mentorship.mentors.template;
-
-enum TemplateStatus {
-    CREATED, CONFIRMED
-}
